+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - Customization for translation files (@goreck888)
 - Customization for scss files (@michal-szostak, @mkasztelnik)
+- Lead section error box displayed to admin user (@martaswiatkowska)
 
 ### Changed
 - Template unification for all modals (@goreck888)